requirements.txt
flask
torch
torchvision
timm
pillow
scikit-learn
opencv-python
numpy
matplotlib
reportlab
pandas
requests
gunicorn
